core: add simplistic back stepping command
Add the "frame_back_step" command (by default bound to ","), which goes
back by one frame. This works by doing precise seeks to the timestamp of
the previously displayed frame. Getting this timestamp is tricky. In
most cases, keeping the timestamps of the N past displayed frames is
sufficient (this commit goes with N=20). But this won't work after seeks
or when trying to step back more than N frames. If this happens, we use
the hr-seek mechanism to index 1 second before the currently displayed
frames. This ought to work with most reasonable video files.

Note that there is one problem: when the user does a normal precise
seek, seeking will try to drop frames with timestamps before the seek
target. These frames won't be added to the timestamp index, and stepping
backward will erroneously skip them. However, this must be causing
trouble for precise seeking as well if certain video filters are
involved, so ignore this issue for now.

In theory, we could incrementally build a more complete index of the
file. But since this is complicated and not very reliable, this doesn't
really seem worth it. It's complicated, because the index would consists
of a set of disjoint indexed runs of already displayed files (if we
really want to build it incrementally). It's unreliable because video
filters and toggling deinterlacing can completely change the set of
frames and their timestamps on the fly. So leave it at this primitive
solution. This also means we add a separate command for this, instead of
trying to integrate it with the "seek" command, or similar.